api_rest/requirements.txt
Django>=3.0,<4.0
djangorestframework
markdown
django-filter
psycopg2-binary>=2.8
django-rest-framework-social-oauth2
drf-yasg
ipdb
django-extensions
gunicorn
whitenoise
dj-database-url